% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLifeTables.R
\name{build_background_mortality_file}
\alias{build_background_mortality_file}
\title{Taking in the CDC NVSS Yearly life tables, extract and build the background mortality table expected by RESPOND}
\usage{
build_background_mortality_file(
  files,
  outputfile,
  races = c("black", "hispanic", "white"),
  sexes = c("female", "male"),
  age_groups = c("1_20", "21_40", "41_60", "61_80", "81_100"),
  bin_size = 20
)
}
\description{
Input: CDC NVSS Life Tables, output file name, stratifications
Output: Resulting background mortality compatible with RESPOND
}
