% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLifeTables.R
\name{build_base_table}
\alias{build_base_table}
\title{Build the base table to fill, stratified by age, sex, and race}
\usage{
build_base_table(
  races = c("black", "hispanic", "white"),
  sexes = c("female", "male"),
  age_groups = c("1_20", "21_40", "41_60", "61_80", "81_100")
)
}
\description{
Input: Demographics to stratify
Output: Empty Shell table
}
\keyword{internal}
