% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLifeTables.R
\name{build_union}
\alias{build_union}
\title{Build the union data.table of various yearly extracted life probabilities}
\usage{
build_union(lt)
}
\arguments{
\item{lt}{A list of data tables containing the weekly probabilities}
}
\description{
Input: List of data.tables containing the weekly probabilities and their stratifications
Output: Joined List of data.tables
}
\keyword{internal}
