% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respond_inputManipulations.R
\name{change_agegrp_chunk}
\alias{change_agegrp_chunk}
\title{A function to change default age groups from RESPOND shell tables}
\usage{
change_agegrp_chunk(data, size.out, transformation, cols, grouping)
}
\arguments{
\item{data}{dataframe or datatable: A RESPOND table}

\item{size.out}{Integer: age-group size out}

\item{transformation}{String: either "mean" or "sum" defining the transformation to new age groups, eg. counts would be summed, transition probabilities would be averaged}

\item{cols}{List: columns to transform}

\item{grouping}{List: columns to group on, the default is any columns not listed in \code{cols}}
}
\description{
A function to change default age groups from RESPOND shell tables
}
