% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{combine_files}
\alias{combine_files}
\title{Combine OUD Count Files into One Data Table}
\usage{
combine_files(path)
}
\arguments{
\item{path}{Character: Path to the directory containing CSV files.}
}
\value{
A \code{data.table} containing all rows from all files, with an added
  \code{source_file} column.
}
\description{
Reads all CSV files in a given directory, merges them into a single data.table,
and adds a \code{source_file} column identifying the origin of each row.
}
