% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_moud.R
\name{compare_moud}
\alias{compare_moud}
\title{Compare MOUD Counts Between 2 Files}
\usage{
compare_moud(old_path, new_path)
}
\arguments{
\item{old_path}{Character: File path to the old MOUD dataset (CSV format)}

\item{new_path}{Character: File path to the new MOUD dataset (CSV format)}
}
\value{
List: A list with two ggplot objects: \code{count_plot} and \code{difference_plot}
}
\description{
This function compares MOUD counts between an "old" and a "new" dataset.
It returns two plots:
(1) Time series comparison by treatment type, and
(2) Difference in counts (new vs old) by treatment type over time.
}
