% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLifeTables.R
\name{create_and_fill_table}
\alias{create_and_fill_table}
\title{Create and fill the table with mortality values for all demographic combinations}
\usage{
create_and_fill_table(
  background_mortality,
  races = c("black", "hispanic", "white"),
  sexes = c("female", "male"),
  age_groups = c("1_20", "21_40", "41_60", "61_80", "81_100")
)
}
\arguments{
\item{background_mortality}{List of extracted background mortality data.tables}

\item{races}{A list of races}

\item{sexes}{A list of sexes}

\item{age_groups}{A list of age groups}
}
\description{
Create and fill the table with mortality values for all demographic combinations
}
\keyword{internal}
