% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respond_input_manipulations.R
\name{DSA}
\alias{DSA}
\title{A Function to Change RESPOND inputs from base-case by some percent change}
\usage{
DSA(data, filter, cycle, pct_change)
}
\arguments{
\item{data}{dataframe: dataframe from RESPOND base input}

\item{filter}{string: treatment block to target the change}

\item{cycle}{list: cycles to edit}

\item{pct_change}{num: percent to change transition state, +3\% -3\% would be 1.03, .97 respectively}
}
\description{
A Function to Change RESPOND inputs from base-case by some percent change
}
