% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLifeTables.R
\name{extract_background_mortality}
\alias{extract_background_mortality}
\title{Function used to extract background mortality values based on age from a single yearly CDC NVSS life table}
\usage{
extract_background_mortality(
  file_path,
  bin_size = 20,
  age_groups = c("1_20", "21_40", "41_60", "61_80", "81_100")
)
}
\description{
Input: CDC Life Tables
Output: data.table containing the weekly probabilities
}
\keyword{internal}
