% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sas.R
\name{fetch_sas}
\alias{fetch_sas}
\title{MA DPH PHDW Script Pull
Fetches most recent .SAS files for MA DPH work and saves them to a SAS/ subdirectory}
\usage{
fetch_sas()
}
\description{
MA DPH PHDW Script Pull
Fetches most recent .SAS files for MA DPH work and saves them to a SAS/ subdirectory
}
