% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_crcTools.R
\name{formula_corr}
\alias{formula_corr}
\title{Internal function to create a formula based on correlation among databases listed in \code{CRC}}
\usage{
formula_corr(corr_matrix, threshold, freq.column)
}
\arguments{
\item{corr_matrix}{A correlation matrix}

\item{threshold}{Threshold for interacting two databases}

\item{freq.column}{frequency column, e.g. N_ID}
}
\description{
Internal function to create a formula based on correlation among databases listed in \code{CRC}
}
\keyword{internal}
