% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{get_age_data}
\alias{get_age_data}
\title{Filter for Age-Specific OUD Data}
\usage{
get_age_data(data)
}
\arguments{
\item{data}{A \code{data.table} containing a \code{source_file} column.}
}
\value{
A filtered \code{data.table} containing only age-related data.
}
\description{
Extracts rows from the dataset corresponding to age-group breakdowns.
}
