% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_cohort_stats.R
\name{get_init_cohort_statistics}
\alias{get_init_cohort_statistics}
\title{Get descriptive initial cohort statistics}
\usage{
get_init_cohort_statistics(db_path, table_name)
}
\arguments{
\item{db_path}{Path where the inputs.db file is saved}

\item{table_name}{Name of the initial cohort table in the
input.db file (e.g., init_cohort, population)}
}
\value{
A table with descriptive statistics of the initial cohort
}
\description{
This function gets descriptive statistics of the initial cohort
for HEP-CE projects. It does this by connecting to an inputs.db file
using SQLite, looking at the init_cohort table, and pulling
descriptive statistics (average age, proportion male, drug behavior
distribution, fibrosis state distribution, etc).
}
