% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{get_monthly_data}
\alias{get_monthly_data}
\title{Filter for Monthly OUD Data}
\usage{
get_monthly_data(data)
}
\arguments{
\item{data}{A \code{data.table} containing a \code{source_file} column.}
}
\value{
A filtered \code{data.table} containing only monthly-related data.
}
\description{
Extracts rows from the dataset corresponding to monthly breakdowns.
}
