% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{get_race_data}
\alias{get_race_data}
\title{Filter for Race-Specific OUD Data}
\usage{
get_race_data(data)
}
\arguments{
\item{data}{A \code{data.table} containing a \code{source_file} column.}
}
\value{
A filtered \code{data.table} containing only race-related data.
}
\description{
Extracts rows from the dataset corresponding to race-based breakdowns.
}
