% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{get_sex_data}
\alias{get_sex_data}
\title{Filter for Sex-Specific OUD Data}
\usage{
get_sex_data(data)
}
\arguments{
\item{data}{A \code{data.table} containing a \code{source_file} column.}
}
\value{
A filtered \code{data.table} containing only sex-related data.
}
\description{
Extracts rows from the dataset corresponding to sex-based breakdowns.
}
