% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good_turing.R
\name{good_turing}
\alias{good_turing}
\title{Good-Turing Estimation of Unseen Species
Provides methodology for the Good-Turing estimation of unseen species for spatial
capture-recapture methodology.}
\usage{
good_turing(data, freq.column, binary.variables)
}
\arguments{
\item{data}{Dataframe: Data for use in the Good-Turing estimation}

\item{freq.column}{Column: A column containing the frequency of observed combinations}

\item{binary.variables}{List of Columns: List containing columns of binary variables indicating involvement in the given database}
}
\description{
Good-Turing Estimation of Unseen Species
Provides methodology for the Good-Turing estimation of unseen species for spatial
capture-recapture methodology.
}
