% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{histogram}
\alias{histogram}
\title{Histogram of OUD Counts}
\usage{
histogram(data)
}
\arguments{
\item{data}{A \code{data.table} containing an \code{N_ID} column.}
}
\value{
A \code{ggplot} object showing the distribution of counts.
}
\description{
Creates a histogram showing the distribution of OUD counts.
}
