% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.formula.R
\name{is.formula}
\alias{is.formula}
\title{Formula Objects
Generic test of an object being interpretable as a formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
Formula Objects
Generic test of an object being interpretable as a formula
}
