% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_github.R
\name{list_files_github}
\alias{list_files_github}
\title{GitHub file listing}
\usage{
list_files_github(user, repo, branch = "main", folder, pattern)
}
\arguments{
\item{user}{string: username}

\item{repo}{string: repository name}

\item{branch}{string: github branch, "main" by default}

\item{folder}{string: reposority subdirectory}

\item{pattern}{string: a pattern of files to match}
}
\description{
GitHub file listing
}
