% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOutputFiles.R
\name{loadOutputFiles}
\alias{loadOutputFiles}
\title{Aggregate selected tables from a set of RESPOND simulation runs}
\usage{
loadOutputFiles(N, outputFileNames, pathPrefix = "")
}
\arguments{
\item{N}{Numeric: The number of output folders to be loaded. Assumes that
output folders are named as \code{output1}, \code{output2}, ...,
\code{outputN}.}

\item{outputFileNames}{Character: A single string or vector of strings
representing the file paths (relative to output#) of files to include in the
returned object.}

\item{pathPrefix}{Character: (Optional) A relative or absolute path that can
be used when loading data not located in the working directory.}
}
\value{
List of Lists: The contents of the specified output folders/files.
The top-level keys are the name of each output folder, i.e. \code{output1},
\code{output2}, etc. These subsequent lists are broken down by the contents of
outputFileNames, e.g.

\if{html}{\out{<div class="sourceCode">}}\preformatted{outputTables
$output1
- $"all_types_overdose1.csv"
- $"general_stats1.csv"
- $"cost_life/total_costs1.csv"
}\if{html}{\out{</div>}}
}
\description{
Aggregate selected tables from a set of RESPOND simulation runs
}
