% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respond_inputManipulations.R
\name{new_block}
\alias{new_block}
\title{A function to create new blocks for RESPOND (clones 'No_Treatment' blocks and reassigns the name)}
\usage{
new_block(data, names)
}
\arguments{
\item{data}{DataFrame: data to manipulate}

\item{names}{Str list: name for the new blocks}
}
\description{
A function to create new blocks for RESPOND (clones 'No_Treatment' blocks and reassigns the name)
}
