% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{plot_age_data}
\alias{plot_age_data}
\title{Plot OUD Counts by Age Group}
\usage{
plot_age_data(age_data)
}
\arguments{
\item{age_data}{A \code{data.table} containing \code{year}, \code{N_ID}, and \code{age_grp_twenty} columns.}
}
\value{
A \code{ggplot} object visualizing counts by age group.
}
\description{
Creates a line plot of OUD counts over time, broken down by age group.
}
