% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{plot_race_data}
\alias{plot_race_data}
\title{Plot OUD Counts by Race}
\usage{
plot_race_data(race_data)
}
\arguments{
\item{race_data}{A \code{data.table} containing \code{year}, \code{N_ID}, and \code{final_re} columns.}
}
\value{
A \code{ggplot} object visualizing counts by race.
}
\description{
Creates a line plot of OUD counts over time, broken down by race.
}
