% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_oud.R
\name{plot_sex_data}
\alias{plot_sex_data}
\title{Plot OUD Counts by Sex}
\usage{
plot_sex_data(sex_data)
}
\arguments{
\item{sex_data}{A \code{data.table} containing \code{year}, \code{N_ID}, and \code{final_sex} columns.}
}
\value{
A \code{ggplot} object visualizing counts by sex.
}
\description{
Creates a line plot of OUD counts over time, broken down by sex.
}
