% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_regression.R
\name{poisson_regression}
\alias{poisson_regression}
\title{Poisson Regression
Regresses \code{data} using input \code{formula} by calling \code{glm} and
and forcing poisson family with a log link function.}
\usage{
poisson_regression(data, formula)
}
\arguments{
\item{data}{Dataframe: input data to the poisson regression model}

\item{formula}{Formula object: formula input into the regression model should only
contain variable names held within \code{data}}
}
\description{
Poisson Regression
Regresses \code{data} using input \code{formula} by calling \code{glm} and
and forcing poisson family with a log link function.
}
