% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_cohort_stats.R
\name{prop_breakdown}
\alias{prop_breakdown}
\title{Break down the initial cohort into proportions}
\usage{
prop_breakdown(var_name, cohort)
}
\arguments{
\item{var_name}{The name of the variable for which to calculate the
proportion breakdown.}

\item{cohort}{The initial cohort data frame (as a dplyr table) from which to
calculate the proportion breakdown.}
}
\value{
A table with the proportion breakdown of the specified value.
}
\description{
This function calculates the proportion breakdown of a given
variable in the initial cohort.
}
\keyword{internal}
