% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respond_input_manipulations.R
\name{replace_vals}
\alias{replace_vals}
\title{A function to change values to another block; follows the format \code{data$x[column == "a", ] <- data$y[column == "b", ]}}
\usage{
replace_vals(data, column = "block", x, y, a, b)
}
\arguments{
\item{data}{DataFrame}

\item{column}{str: column name}

\item{x}{str}

\item{y}{str}

\item{a}{str}

\item{b}{str}
}
\description{
A function to change values to another block; follows the format \code{data$x[column == "a", ] <- data$y[column == "b", ]}
}
