% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{separate}
\alias{separate}
\title{User-friendly split: Separates a dataframe into a list of dataframes along a specified column with an extra option to write out the data}
\usage{
separate(data, column.name, save = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{column.name}{variable: column to separate upon}

\item{save}{boolean: by default FALSE indicating not to write to csv}
}
\description{
User-friendly split: Separates a dataframe into a list of dataframes along a specified column with an extra option to write out the data
}
