% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syndemics-package.R
\docType{package}
\name{Syndemics-package}
\alias{Syndemics}
\alias{Syndemics-package}
\title{Syndemics: A collection of miscellaneous functions created for internal use with the Boston Medical Center Syndemics lab}
\description{
The Syndemics Lab is a research unit with Boston Medical Center. Our primary work involves dynamic simulations for predicting and analyzing opioid use disorder, HIV, and HEP-CE. This package is a collection of commonly used workflow functionalizations serving analytic purposes.
}
\keyword{internal}
