% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SYNDEMICS_BLUE}
\alias{SYNDEMICS_BLUE}
\title{Syndemics Branding Color - Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
SYNDEMICS_BLUE
}
\description{
The dark blue used in Syndemics Lab-branded products
}
\keyword{datasets}
