% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SYNDEMICS_COLORS}
\alias{SYNDEMICS_COLORS}
\title{All Syndemics Branding Colors}
\format{
An object of class \code{character} of length 3.
}
\usage{
SYNDEMICS_COLORS
}
\description{
A vector of all three Syndemics Lab colors, provided for
convenience for use in plots and other visuals
}
\keyword{datasets}
