% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SYNDEMICS_CYAN}
\alias{SYNDEMICS_CYAN}
\title{Syndemics Branding Color - Cyan}
\format{
An object of class \code{character} of length 1.
}
\usage{
SYNDEMICS_CYAN
}
\description{
The cyan used in Syndemics Lab-branded products
}
\keyword{datasets}
