% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SYNDEMICS_PINK}
\alias{SYNDEMICS_PINK}
\title{Syndemics Branding Color - Pink}
\format{
An object of class \code{character} of length 1.
}
\usage{
SYNDEMICS_PINK
}
\description{
The pink used in Syndemics Lab-branded products
}
\keyword{datasets}
