% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc.R
\name{crc}
\alias{crc}
\title{Spatial Capture Re-Capture}
\usage{
crc(
  data,
  freq.column,
  binary.variables,
  method = "poisson",
  formula.selection = "stepwise",
  corr.threshold = 0.2,
  formula = NULL,
  opts.stepwise = list(direction = "both", threshold = 0.05, verbose = TRUE)
)
}
\arguments{
\item{data}{Dataframe: A dataframe containing a frequency column and binary columns indicating involvement in the given database}

\item{freq.column}{Column: A column containing the frequency of observed combinations}

\item{binary.variables}{List of Columns: List containing columns of binary variables indicating involvement in the given database}

\item{method}{String: Selection for the spatial capture-recapture method - either poisson, or negbin}

\item{formula.selection}{String: Selection for formula decision when \code{method} is poisson or negbin - either aic, corr, or stepwise}

\item{corr.threshold}{Numeric: Threshold for forcing interaction term between binary columns. Only applicable when \code{formula.selection} is \code{"corr"}}

\item{formula}{Formula: Allows definition of custom formula object for poisson regression. In the case of a specified formula, both \code{formula.selection} methods will produce the same results}

\item{opts.stepwise}{List: List of \code{direction}: 'forward', 'backward', or 'both', \code{threshold}: p-value threshold for stepwise selection, and \code{verbose} if you would like every stepped-through model to be printed}
}
\description{
A method for estimation of 'unknowns' through knowledge about knowns as described by
Barocas, Joshua A et al. “Estimated Prevalence of Opioid Use Disorder in Massachusetts, 2011-2015: A Capture-Recapture Analysis.” doi:10.2105/AJPH.2018.304673
\code{crc} now builds on other functions inside this package: \code{corr_formula} builds a correlation matrix between \code{binary.variables} using \code{freq.column}
then constructs a recommended formula for poisson regression models. If the correlation is greater than \code{corr.threshold} it assumes an interaction term instead of additive.
There are options to use stepwise selection which step-wise selects models based on \code{opts.stepwise}'s \code{threshold} and \code{DBCount}, which
rowwise sums the \code{binary.variables}, sums \code{freq.column} by grouping the rowwise sum, and uses poisson regression to estimate \code{freq.column ~ DBCount}
}
