% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sas.R
\name{fetch_sas}
\alias{fetch_sas}
\title{Fetches most recent .SAS files for MA DPH work and saves them to a SAS/ subdirectory}
\usage{
fetch_sas()
}
\description{
Fetches most recent .SAS files for MA DPH work and saves them to a SAS/ subdirectory
}
