% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLifeTables.R
\name{fill_base_table}
\alias{fill_base_table}
\title{Fill the shell table with values extracted}
\usage{
fill_base_table(
  base_table,
  background_mortality,
  races = c("black", "hispanic", "white"),
  sexes = c("female", "male")
)
}
\description{
Input: Empty shell table and extracted background mortality
Output: a full data.table containing the extracted mortalities
}
\keyword{internal}
