% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syndemics-package.R
\docType{package}
\name{Syndemics-package}
\alias{Syndemics}
\alias{Syndemics-package}
\title{Syndemics: Syndemics Lab at Boston Medical Center Tools for Interacting with Simulation Models}
\description{
Tools for dynamic simulation modeling and policy analysis of opioid use disorder, HIV, and Hepatitis C epidemics. This package provides workflow functions for interacting with RESPOND simulation models, enabling users to perform simulated costing analysis of different policy interventions and public health measures. The baseline primary function implements an extended Capture-Recapture prevalence estimation method as described in Barocas et al. (2018) doi:10.2105/AJPH.2018.304673, accessible through the crc function, which estimates total background population. Additional functions support input preparation and output analysis for dynamic epidemic simulations. RESPOND models and methodology details available at \url{https://www.syndemicslab.org/respond}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SyndemicsLab/Syndemics}
  \item Report bugs at \url{https://github.com/SyndemicsLab/Syndemics/issues}
}

}
\author{
\strong{Maintainer}: Matthew Carroll \email{matthew.carroll@bmc.org}

Authors:
\itemize{
  \item Ryan O'Dea \email{ryan.odea@bmc.org}
  \item Dimitri Baptiste \email{dimitri.baptiste@bmc.org}
  \item Hana Zwick \email{hana.zwick@bmc.org}
}

}
\keyword{internal}
